Added scripts to take care of simon's data
Still there are some discrepancies in the format between
the data from simon and the forma used (csv).
The plan is to get rid of CSV altogether, while adding a header
to the file "as is".
(CSV is not needed for dataframe if read_table and sep='\s+', or
something like that, is used instead of read_csv).