Improved utility for checking thermalisation and block size
Now the lib.scan_for_blocking function plots the data and asks for
optional new values.
If 'return' is pressed, the old value is used.
If 'N' is entered, the parameter is 'flagged' as "need more statistics"
and it is written into a file for convenience.

At the end, a new "analysis settings" file is outputted, with the new
parameters in place of the old ones (if they were changed).